requirements.txt
delta-spark==2.1.0
importlib_metadata==7.1.0
py4j==0.10.9.5
pyspark==3.3.0
zipp==3.19.2
